requirements.txt
# Core Dependencies
streamlit>=1.31.0
pandas>=2.1.4
numpy>=1.26.3

# LLM and Multi-Agent Framework
openai>=1.12.0
google-generativeai>=0.3.0
langchain>=0.1.7
langchain-openai>=0.0.5
langchain-google-genai>=0.0.5
langgraph>=0.0.20
langchain-community>=0.0.20

# Data Processing and Storage
duckdb>=0.9.2
pyarrow>=15.0.0
openpyxl>=3.1.2

# Vector Store (Optional but included)
faiss-cpu>=1.7.4
chromadb>=0.4.22

# Visualization and UI
plotly>=5.18.0
matplotlib>=3.8.2
seaborn>=0.13.1

# Utilities
python-dotenv>=1.0.1
pydantic>=2.5.3
tiktoken>=0.5.2

